Retire arbitrary student captain "group"
The concepts of student captains are something left over from the old
days of the FOSSBox. For as long as I have been a student at RIT, I have
never seen this term used to describe anyone. I think an arbitrary
separation based on unclear selection criteria is an unnecessary point
of division in what is really closer to equality.