:alembic: ci: Add Travis config and Rakefile with html-proofer
This adds the same CI tests that the FOSSRIT website uses. Since this
repo supports the `gh-pages` branch, I will probably send up a future PR
that sets up gated deployments that are conditional on whether tests are
passing in `master`.

For now, this is a start!

Also worth noting: YAML files are also linted for valid YAML. The config
file also had some changes and better organization in order to make sure
Travis builds ignore the `vendor/` directory.

Signed-off-by: Justin W. Flory <git@jwf.io>