:building_construction: ci: Deploy built site from Travis (#84)
Since I already did all this work, might as well do the last bit. This
commit adds a gated deployment from Travis. Nothing changes to the test
suite, but now, when the tests pass on `master`, the generated site
source in `_site/` will be pushed to the `gh-pages` branch. Once this
merges, I'll switch the repo to publish the content in that branch.

It's one way to make sure that if the tests ever fail and things get
messy in `master`, the bad content doesn't build unless tests actually
pass.

Signed-off-by: Justin W. Flory <git@jwf.io>