:question: config: Expand excludes, set permalink to pretty
This is a slight modification to the site config while I am debugging
the wider issue of why the site isn't rendering correctly on all pages.
For now, I am making sure that excluded files are correctly ignored and
I enabled the `pretty` setting for `permalink` to trim off the excess
`.html` on URLs.

Signed-off-by: Justin W. Flory <git@jwf.io>