:wrench: config: Set baseurl to `people` for CSS to render (#86)
This fixes a bug by setting the `baseurl` setting in the config. We need
this since the site is served under a subdirectory of the main domain,
`fossrit.github.io`. The CSS should start rendering normally again after
this. My local test environment worked.

I had to disable external link checking on `htmlproofer` though. Not
happy about it, but I need some more clever way of fixing this.
`url_swap`?

Signed-off-by: Justin W. Flory <git@jwf.io>