:mailbox_with_mail: workflows: Add inclusive-organization-action@v1.1.0 (#91)
This GitHub Action will automatically send an invite to someone who
contributes a merged pull request to FOSSRIT/people. It invites them
specifically to @FOSSRIT/friends-of-foss, which is a view-only access
team. If they need a greater level of access, a human needs to add them
to another team manually.

This can be part of a FOSSbox on-boarding platform, in a way. :smile:

Signed-off-by: Justin W. Flory <git@jwf.io>