:alembic: :wrench: ci: Migrate from Travis CI to Circle CI
Lately, Travis CI jobs take forever to finish. Furthermore, their
pricing plans have become more hostile for Open Source and builds that
take a long time to complete. Which is hard to avoid, when it is their
platform that is causing the builds to be so slow…

So, this is a mostly invisible commit to migrate the CI over to Circle
CI. No noticeable changes except that the site is now being built with
Ruby v3.0.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>