:beetle: :wrench: ci: Define jobs in Circle CI config
So, I was too used to using orbs where a job is already pre-defined.
Since I am not using Circle CI Orbs here, I have to define the job in
the config file itself. I cannot define the steps of a job within a
workflow. Which, I mean, I guess it makes sense.

So, now, this commit defines the jobs so Circle CI should at least
attempt to try them out for size.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>