:package: dependencies: Periodic dependency update, patch multiple CVEs
This commit is the result of updating gem dependencies, and the addition
of a new explicit dependency in order to patch a CVE. It is not ideal to
hard-code the `rack` version in the Gemfile, but it fixes a CVE and
GitHub will finally stop emailing me if I do this.

Also, this should force the Circle CI cache to refresh since the SHA256
sum of the `Gemfile.lock` changed. So, it should use the correct paths
for future builds (assuming this one is successful).

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>