:boom: Purge old Python-based Jinja site from repo
GitHub's dependency updates bot doesn't realize the Python dependency
files here are not in use and are part of the old site's archive.
However, since this is git, the old site will always live on in the git
history of this repo. So, to reduce unnecessary noise and panic by the
GitHub dependency bot, this commit removes the old site from the repo.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>